requirements.txt
fastapi
pydantic
requests
pytest
pytest-cov
python-dateutil